refactor: move search bar to components folder
